Support of "caps" tag
This patch allows the support of a "caps" tag, that defines the value of the key when the CAPS action modifier is in effect.
This feature is needed for the french keyboard - at least - in order to handle the uppercase accented letters, that are the ones accessible when caps is on, for some keys.
There is no impact when the tag is not used in the xml description of the layout, therefore there should be no impact on keyboards not using this tag.
Tests have been done using sato.

Signed-off-by: Herve Jourdain <herve.jourdain@neuf.fr>
Signed-off-by: Jussi Kukkonen <jussi.kukkonen@intel.com>